create pages directory, add simple example to page, update app.js
